fix(lab): preserve fractional ages in native reference selection

+++ native/lab/include/medicore/lab/deterministic_core.hpp
@@ -35,8 +35,8 @@ struct ReferenceCandidate {
     std::string unit;
     std::string source;
     std::string sex{"ANY"};
-    std::optional<int> age_min;
-    std::optional<int> age_max;
+    std::optional<double> age_min;
+    std::optional<double> age_max;
     std::optional<bool> pregnancy_status;
 };
 
@@ -48,10 +48,7 @@ struct ReferenceSelection {
     std::string reason;
 };
 
-// Canonical deterministic normalization for lab names and aliases.
 std::string normalize_alias(const std::string& value);
-
-// Python SequenceMatcher-like Gestalt similarity for short normalized lab names.
 double alias_similarity_ratio(const std::string& left, const std::string& right);
 
 RuleEvaluation evaluate_rule(
@@ -71,7 +68,7 @@ TrendEvaluation compare_trend(
 ReferenceSelection select_reference_candidate(
     const std::vector<ReferenceCandidate>& candidates,
     const std::string& patient_sex,
-    const std::optional<int>& patient_age,
+    const std::optional<double>& patient_age,
     const std::optional<bool>& pregnancy_status);
 
 }  // namespace medicore::lab
